Split association serializers in several files
