+++ associations/serializers/__init__.py
@@ -1,0 +1,5 @@
+from associations.serializers.association import *
+from associations.serializers.library import *
+from associations.serializers.marketplace import *
+from associations.serializers.news import *
+from associations.serializers.page import *
