+++ associations/serializers/library.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from associations.models import Library
+
+
+class LibrarySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Library
+        fields = ('id', 'enabled')
