+++ associations/serializers/news.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from associations.models import News
+
+
+class NewsSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = News
+        fields = ("id", "title", "text", "association")
