+++ associations/serializers/page.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+
+from associations.models import Page
+
+
+class PageSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Page
+        fields = ("id", "title", "text", "association")
+
+
+class PageShortSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Page
+        fields = ("id", "title")
